Реализовал запуск многопоточного обхода страниц в отдельных потоках.

+++ src/main/java/searchengine/services/IndexServiceImpl.java
@@ -4,18 +4,19 @@ import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
 import searchengine.config.Site;
 import searchengine.config.SitesList;
+import searchengine.dto.IndexResponse;
 import searchengine.dto.NodePage;
-import searchengine.dto.ResponseDto;
 import searchengine.entity.SiteEntity;
 import searchengine.entity.enumerated.StatusType;
 import searchengine.repository.PageRepository;
 import searchengine.repository.SiteRepository;
 import searchengine.services.interfaces.IndexService;
-import searchengine.webCrawler.RecursiveWebCrawler;
+import searchengine.webCrawler.MultithreadedWebCrawler;
 import searchengine.webCrawler.interfaces.PageParser;
 
 import java.time.LocalDateTime;
-import java.util.Map;
+import java.util.ArrayList;
+import java.util.List;
 import java.util.concurrent.ForkJoinPool;
 
 @RequiredArgsConstructor
@@ -26,37 +27,42 @@ public class IndexServiceImpl implements IndexService {
     private final PageRepository pageRepository;
     private final PageParser pageParser;
     private final SitesList sites;
-    private final ForkJoinPool forkJoinPool;
-
-
-    /*
-     * TODO:
-     *  1. Придумать как в startIndexing реализовать верно обработку листа сайтов
-     *  2. Протестировать!
-     * */
+    private final ForkJoinPool forkJoinPool = new ForkJoinPool();
 
     @Override
-    public ResponseDto<Map<String, Boolean>> startIndexing() {
-        var site = sites.getSites().get(0);
+    public IndexResponse startIndexing() {
+        var sitesList = sites.getSites();
+        List<NodePage> nodePages = new ArrayList<>();
 
-        deleteAllInfoFromDataBase(site);
-        var siteId = saveSiteInDataBaseAndReturnId(site);
+        for (Site site : sitesList) {
+            var isSiteExist = siteRepository.existsByUrl(site.getUrl());
+            var status = isSiteExist ? siteRepository.findByUrl(site.getUrl()).getStatus() : StatusType.INDEXED;
 
-        NodePage nodePage = new NodePage();
-        nodePage.setPath(site.getUrl())
-                .setSuffix(site.getUrl())
-                .setPrefix("")
-                .setTimeBetweenRequest(150)
-                .setSiteId(siteId);
+            if (status.equals(StatusType.INDEXING)) return new IndexResponse()
+                    .setResult(false)
+                    .setError("Индексация уже запущена");
 
-        forkJoinPool.invoke(new RecursiveWebCrawler(pageParser, nodePage));
+            var siteEntity = getSiteEntity(site);
 
-        var status = siteRepository.findByUrl(site.getUrl()).getStatus();
-        return ResponseDto.<Map<String, Boolean>>builder()
-                .data(Map.of("result", !status.equals(StatusType.INDEXING)))
-                .error(status.equals(StatusType.INDEXING) ? "Индексация  уже запущена" : "")
-                .build();
+            if (isSiteExist) deleteAllInfoFromDataBase(site);
+            saveSiteInDataBase(siteEntity);
 
+            var nodePage = getNodePage(site);
+            nodePages.add(nodePage);
+        }
+
+        nodePages.forEach(nodePage -> new MultithreadedWebCrawler(forkJoinPool, pageParser, nodePage).start());
+        setStatusIndexed();
+
+        return new IndexResponse().setResult(true);
+    }
+
+    private SiteEntity getSiteEntity(Site site) {
+        return new SiteEntity().setStatus(StatusType.INDEXING)
+                .setStatusTime(LocalDateTime.now())
+                .setLastError(null)
+                .setUrl(site.getUrl())
+                .setName(site.getName());
     }
 
     private void deleteAllInfoFromDataBase(Site site) {
@@ -67,29 +73,40 @@ public class IndexServiceImpl implements IndexService {
         }
     }
 
-    private int saveSiteInDataBaseAndReturnId(Site site) {
-        return siteRepository.save(new SiteEntity().setStatus(StatusType.INDEXING)
-                        .setStatusTime(LocalDateTime.now())
-                        .setLastError(null)
-                        .setUrl(site.getUrl())
-                        .setName(site.getName()))
-                .getId();
+    private void saveSiteInDataBase(SiteEntity siteEntity) {
+        if (siteEntity != null) siteRepository.save(siteEntity);
+    }
+
+    private NodePage getNodePage(Site site) {
+        var siteEntity = siteRepository.findByUrl(site.getUrl());
+        return new NodePage().setPath(site.getUrl())
+                .setSuffix(site.getUrl())
+                .setPrefix("")
+                .setTimeBetweenRequest(150)
+                .setSiteId(siteEntity.getId());
+    }
+
+    private void setStatusIndexed() {
+        var indexingSites = siteRepository.findAllByStatus(StatusType.INDEXING);
+        indexingSites.forEach(site -> site.setStatus(StatusType.INDEXED));
+        siteRepository.saveAll(indexingSites);
     }
 
     @Override
-    public ResponseDto<Map<String, Boolean>> stopIndexing() {
-        boolean isIndexing;
-
+    public IndexResponse stopIndexing() {
         if (forkJoinPool.isShutdown()) {
-            isIndexing = false;
+            return new IndexResponse().setResult(false).setError("Индексация не запущена");
         } else {
-            isIndexing = true;
             forkJoinPool.shutdown();
+            setStatusFailedForNotIndexedSites();
         }
 
-        return ResponseDto.<Map<String, Boolean>>builder()
-                .data(Map.of("result", isIndexing))
-                .error(isIndexing ? "" : "Индексация не запущена")
-                .build();
+        return new IndexResponse().setResult(true);
+    }
+
+    private void setStatusFailedForNotIndexedSites() {
+        var notIndexedSites = siteRepository.findAllByStatus(StatusType.INDEXING);
+        notIndexedSites.forEach(site -> site.setStatus(StatusType.FAILED).setLastError("Индексация остановлена пользователем"));
+        siteRepository.saveAll(notIndexedSites);
     }
 }
